Making updater/Android.mk more compilant with other tools
Moved commands generating ".list" file to ".list" file target
from "hackish" variable definition. Semantics stays roughly the same - we
will always try to remake file, it will be updated only if necessary.
The only difference is that it will be done during build stage, not
Makefile reading stage (which would be more correct). Both solutions
work under GNU Make.

The main motive for the change is usage of other tools (e.g. dependency
view tool). On clean source tree it will fail, because there is no way
to remake ".list" target (file does not exist and there are no rules to
build it) which is a prerequisite to "$(inc)" file.

It might be a cosmetic change, but it's the only place I've found in
android, where a file in "out" directory is made during Makefile read
stage.

Change-Id: I52936b618736a66f8198a1cecc39589435aa4db5
Signed-off-by: Marek Białowąs <nalajcie@gmail.com>