minui: allow device to skip the initial display blank
The theory behind the blank was that the kernel would not
re-initilize the display on boot (and leave it to whatever
the bootloader configured it to) until after the first
blank request.

On some devices, the bootloader does not turn the screen on,
but it's done by the kernel and thus. This makes the extra blank at
start unnecessary

Bug: 7492696
Change-Id: I1af9bc07c15f1660846192467378f3a1585dad3a