Enable wrap-around behavior for recovery menu items.
The current code requires that a device has at least 3 buttons or
button combos: up, down and select.

When running recovery on devices with less than 3 hardware buttons,
it is useful to allow the menu selection to wrap back to the start
when the end of the menu is reached.

Change-Id: Iebd037930667d80f719a2aa6d226d36797c41541