recovery: Properly detect userdebug or eng builds
The recovery system behaves a little bit different on userdebug or
eng builds by presenting error reports to the user in the ui.
This is controlled by checking the build fingerprint for the string
:userdebug/ or :eng/. But with AOSP version numbers most AOSP
builds blows the 92 char limit of ro.build.fingerprint and therefor
the property is not set, so this condition will always be evaluated
to false, for most builds.

Instead of depending on the flaky ro.build.fingerprint this change
makes the code read the specific part directly of the fingerprint
directly from ro.build.type.

Change-Id: I74bc00c655ac596aaf4b488ecea58f0a8de9c26b